requirements.txt
torch
torchvision
datasets
trl==0.15.2
qwen_vl_utils
mathruler
pylatexenc
Pillow
